Ported the high-level IO/storage classes over from AtheOS.
- Ported the high-level IO/storage classes over from AtheOS.
- Added versions of Open(), CreateDirectory(), Unlink() and RemoveDirectory() that accept a directory file-descriptor to be used as the base for the path when looking up a file or directory (rather than basing it off the current working directory).
- Added an API for doing a deep copy of a file descriptor. Similar to dupe(), but creates a new file object with it's own state like current position but referencing the same file.
- Added an API for back-traversing from an open directory to build the primary path leading to the directory.
- KFileTableNode now tracks the flags used when opening the file.

+++ System/GUI/ListViewHeaderView.cpp
@@ -79,7 +79,7 @@ void ListViewHeaderView::Paint(const Rect& cUpdateRect)
     FontHeight sHeight = pcFont->GetHeight();
 
     Rect cFrame;
-    for (uint i = 0; i < m_ScrolledContainerView->m_ColumnMap.size(); ++i)
+    for (size_t i = 0; i < m_ScrolledContainerView->m_ColumnMap.size(); ++i)
     {
         Ptr<ListViewColumnView> pcCol = m_ScrolledContainerView->GetColumn(i);
         cFrame = pcCol->GetFrame();
@@ -240,7 +240,7 @@ bool ListViewHeaderView::OnMouseUp(MouseButton_e button, const Point& position, 
     //      int column;
     //      if (pcData->FindInt("column", &column) == 0) {
     //          Point cMVPos = m_pcMainView->ConvertFromParent(position);
-    //          for (uint i = 0; i < m_pcMainView->m_ColumnMap.size(); ++i)
+    //          for (size_t i = 0; i < m_pcMainView->m_ColumnMap.size(); ++i)
     //          {
     //              Rect cFrame(m_pcMainView->GetColumn(i)->GetFrame());
     //
